+++ src/completion/source-language.ts
@@ -108,6 +108,7 @@ export default class LanguageSource implements ISource<CompletionItem> {
       } catch (e) {
         invalid = true
         this.resolving.delete(item)
+        // eslint-disable-next-line @typescript-eslint/prefer-promise-reject-errors
         reject(e)
       }
     })
@@ -119,7 +120,7 @@ export default class LanguageSource implements ISource<CompletionItem> {
 
   public async onCompleteDone(item: CompletionItem, opt: CompleteDoneOption): Promise<void> {
     let doc = workspace.getDocument(opt.bufnr)
-    await doc.patchChange(true)
+    await doc.patchChange()
     let additionalEdits = !isFalsyOrEmpty(item.additionalTextEdits)
     let version = doc.version
     let isSnippet = await this.applyTextEdit(doc, additionalEdits, item, opt)