+++ src/handler/format.ts
@@ -7,10 +7,9 @@ import events from '../events'
 import languages, { ProviderName } from '../languages'
 import { createLogger } from '../logger'
 import Document from '../model/document'
-import snippetManager from '../snippets/manager'
 import { IConfigurationChangeEvent } from '../types'
 import { isFalsyOrEmpty } from '../util/array'
-import { defaultValue, pariedCharacters } from '../util/index'
+import { pariedCharacters } from '../util/index'
 import { CancellationTokenSource } from '../util/protocol'
 import { isAlphabet } from '../util/string'
 import window from '../window'
@@ -57,6 +56,7 @@ export default class FormatHandler {
           const provideEdits = languages.provideDocumentFormattingEdits(event.document, options, tokenSource.token)
           let textEdits = await Promise.race([tp, provideEdits])
           clearTimeout(timer)
+          this.logProvider(event.bufnr, textEdits)
           return Array.isArray(textEdits) ? textEdits : undefined
         }
         event.waitUntil(willSaveWaitUntil())
@@ -124,7 +124,9 @@ export default class FormatHandler {
       await doc.synchronize()
       return await languages.provideDocumentOnTypeEdits(ch, doc.textDocument, position, token)
     })
-    await doc.applyEdits(defaultValue(edits, []), false, true)
+    if (edits.length === 0) return true
+    await doc.applyEdits(edits, false, true)
+    this.logProvider(doc.bufnr, edits)
     return true
   }
 
@@ -149,6 +151,7 @@ export default class FormatHandler {
     })
     if (textEdits && textEdits.length > 0) {
       await doc.applyEdits(textEdits, false, true)
+      this.logProvider(doc.bufnr, textEdits)
       return true
     }
     return false
@@ -162,7 +165,7 @@ export default class FormatHandler {
     await this.tryFormatOnType('\n', doc)
     if (bracketEnterImprove) {
       let line = (await nvim.call('line', '.') as number) - 1
-      await doc._fetchContent(false)
+      await doc.patchChange()
       let pre = doc.getline(line - 1)
       let curr = doc.getline(line)
       let firstLine = doc.getline(0)
@@ -191,6 +194,12 @@ export default class FormatHandler {
     }
   }
 
+  public logProvider(bufnr: number, edits: TextEdit[] | undefined): void {
+    if (!Array.isArray(edits) || edits.length === 0) return
+    let extensionName = edits['__extensionName']
+    if (extensionName) logger.info(`Format buffer ${bufnr} by ${extensionName}`)
+  }
+
   public async documentRangeFormat(doc: Document, mode?: string): Promise<number> {
     this.handler.checkProvider(ProviderName.FormatRange, doc.textDocument)
     await doc.synchronize()
@@ -210,6 +219,7 @@ export default class FormatHandler {
     })
     if (!isFalsyOrEmpty(textEdits)) {
       await doc.applyEdits(textEdits, false, true)
+      this.logProvider(doc.bufnr, textEdits)
       return 0
     }
     return -1