+++ src/language-client/index.ts
@@ -1,6 +1,6 @@
 'use strict'
 /* eslint-disable no-redeclare */
-import { ChildProcess, ChildProcessWithoutNullStreams, ForkOptions as CForkOptions } from 'child_process'
+import { ForkOptions as CForkOptions, ChildProcess, ChildProcessWithoutNullStreams } from 'child_process'
 import { createLogger } from '../logger'
 import { disposeAll, getConditionValue } from '../util'
 import * as Is from '../util/is'
@@ -303,8 +303,10 @@ export class LanguageClient extends BaseLanguageClient {
     if (runDebug.run || runDebug.debug) {
       if (typeof v8debug === 'object' || this._forceDebug || startedInDebugMode(process.execArgv)) {
         json = runDebug.debug
+        this._isInDebugMode = true
       } else {
         json = runDebug.run
+        this._isInDebugMode = false
       }
     } else {
       json = server as NodeModule | Executable
@@ -396,7 +398,7 @@ export class LanguageClient extends BaseLanguageClient {
         this._isDetached = !!options.detached
         return Promise.resolve({ reader: new StreamMessageReader(serverProcess.stdout), writer: new StreamMessageWriter(serverProcess.stdin) })
       }
-      return Promise.reject<MessageTransports>(`Unsupported server configuration ${JSON.stringify(server, null, 2)}`)
+      return Promise.reject<MessageTransports>(new Error(`Unsupported server configuration ${JSON.stringify(server, null, 2)}`))
     })
   }