+++ src/util/map.ts
@@ -6,11 +6,16 @@ interface Item<K, V> {
   value: V
 }
 
-export const enum Touch {
-  None = 0,
-  AsOld = 1,
-  AsNew = 2
+export namespace Touch {
+  export const None = 0
+  export const First = 1
+  export const AsOld = First
+  export const Last = 2
+  export const AsNew: 2 = Last
 }
+
+// eslint-disable-next-line no-redeclare
+export type Touch = 0 | 1 | 2
 
 export class LinkedMap<K, V> implements Map<K, V> {
 
@@ -23,7 +28,7 @@ export class LinkedMap<K, V> implements Map<K, V> {
 
   private _state: number
 
-  constructor() {
+  public constructor() {
     this._map = new Map<K, Item<K, V>>()
     this._head = undefined
     this._tail = undefined
@@ -55,6 +60,16 @@ export class LinkedMap<K, V> implements Map<K, V> {
     return this._tail?.value
   }
 
+  public before(key: K): V | undefined {
+    const item = this._map.get(key)
+    return item ? item.previous?.value : undefined
+  }
+
+  public after(key: K): V | undefined {
+    const item = this._map.get(key)
+    return item ? item.next?.value : undefined
+  }
+
   public has(key: K): boolean {
     return this._map.has(key)
   }
@@ -83,10 +98,10 @@ export class LinkedMap<K, V> implements Map<K, V> {
         case Touch.None:
           this.addItemLast(item)
           break
-        case Touch.AsOld:
+        case Touch.First:
           this.addItemFirst(item)
           break
-        case Touch.AsNew:
+        case Touch.Last:
           this.addItemLast(item)
           break
         default:
@@ -142,16 +157,14 @@ export class LinkedMap<K, V> implements Map<K, V> {
   }
 
   public keys(): IterableIterator<K> {
-    // eslint-disable-next-line @typescript-eslint/no-this-alias
-    const map = this
     const state = this._state
     let current = this._head
     const iterator: IterableIterator<K> = {
-      [Symbol.iterator]() {
+      [Symbol.iterator]: () => {
         return iterator
       },
-      next(): IteratorResult<K> {
-        if (map._state !== state) {
+      next: (): IteratorResult<K> => {
+        if (this._state !== state) {
           throw new Error(`LinkedMap got modified during iteration.`)
         }
         if (current) {
@@ -167,16 +180,14 @@ export class LinkedMap<K, V> implements Map<K, V> {
   }
 
   public values(): IterableIterator<V> {
-    // eslint-disable-next-line @typescript-eslint/no-this-alias
-    const map = this
     const state = this._state
     let current = this._head
     const iterator: IterableIterator<V> = {
-      [Symbol.iterator]() {
+      [Symbol.iterator]: () => {
         return iterator
       },
-      next(): IteratorResult<V> {
-        if (map._state !== state) {
+      next: (): IteratorResult<V> => {
+        if (this._state !== state) {
           throw new Error(`LinkedMap got modified during iteration.`)
         }
         if (current) {
@@ -192,16 +203,14 @@ export class LinkedMap<K, V> implements Map<K, V> {
   }
 
   public entries(): IterableIterator<[K, V]> {
-    // eslint-disable-next-line @typescript-eslint/no-this-alias
-    const map = this
     const state = this._state
     let current = this._head
     const iterator: IterableIterator<[K, V]> = {
-      [Symbol.iterator]() {
+      [Symbol.iterator]: () => {
         return iterator
       },
-      next(): IteratorResult<[K, V]> {
-        if (map._state !== state) {
+      next: (): IteratorResult<[K, V]> => {
+        if (this._state !== state) {
           throw new Error(`LinkedMap got modified during iteration.`)
         }
         if (current) {
@@ -220,7 +229,10 @@ export class LinkedMap<K, V> implements Map<K, V> {
     return this.entries()
   }
 
-  public trimOld(newSize: number) {
+  public trimOld(newSize: number): void {
+    if (newSize >= this.size) {
+      return
+    }
     if (newSize === 0) {
       this.clear()
       return
@@ -289,11 +301,11 @@ export class LinkedMap<K, V> implements Map<K, V> {
   }
 
   private touch(item: Item<K, V>, touch: Touch): void {
-    if ((touch !== Touch.AsOld && touch !== Touch.AsNew)) {
+    if ((touch !== Touch.First && touch !== Touch.Last)) {
       return
     }
 
-    if (touch === Touch.AsOld) {
+    if (touch === Touch.First) {
       if (item === this._head) {
         return
       }
@@ -320,7 +332,7 @@ export class LinkedMap<K, V> implements Map<K, V> {
       this._head.previous = item
       this._head = item
       this._state++
-    } else {
+    } else if (touch === Touch.Last) {
       if (item === this._tail) {
         return
       }
@@ -371,7 +383,7 @@ export class LRUCache<K, V> extends LinkedMap<K, V> {
   private _limit: number
   private _ratio: number
 
-  constructor(limit: number, ratio = 1) {
+  public constructor(limit: number, ratio = 1) {
     super()
     this._limit = limit
     this._ratio = Math.min(Math.max(0, ratio), 1)
@@ -395,7 +407,7 @@ export class LRUCache<K, V> extends LinkedMap<K, V> {
     this.checkTrim()
   }
 
-  public override get(key: K, touch: Touch = Touch.AsNew): V | undefined {
+  public get(key: K, touch: Touch = Touch.AsNew): V | undefined {
     return super.get(key, touch)
   }
 
@@ -403,8 +415,8 @@ export class LRUCache<K, V> extends LinkedMap<K, V> {
     return super.get(key, Touch.None)
   }
 
-  public override set(key: K, value: V): this {
-    super.set(key, value, Touch.AsNew)
+  public set(key: K, value: V): this {
+    super.set(key, value, Touch.Last)
     this.checkTrim()
     return this
   }