+++ src/workspace.ts
@@ -1,6 +1,6 @@
 'use strict'
 import { Neovim } from '@chemzqm/neovim'
-import type { DocumentSelector, WorkspaceFoldersChangeEvent } from 'vscode-languageserver-protocol'
+import type { DocumentFilter, DocumentSelector, WorkspaceFoldersChangeEvent } from 'vscode-languageserver-protocol'
 import { TextDocument } from 'vscode-languageserver-textdocument'
 import { CreateFileOptions, DeleteFileOptions, FormattingOptions, Location, LocationLink, Position, Range, RenameFileOptions, WorkspaceEdit, WorkspaceFolder } from 'vscode-languageserver-types'
 import { URI } from 'vscode-uri'
@@ -15,7 +15,7 @@ import Editors from './core/editors'
 import { FileSystemWatcher, FileSystemWatcherManager } from './core/fileSystemWatcher'
 import Files, { FileCreateEvent, FileDeleteEvent, FileRenameEvent, FileWillCreateEvent, FileWillDeleteEvent, FileWillRenameEvent, TextDocumentWillSaveEvent } from './core/files'
 import { callAsync, createNameSpace, findUp, getWatchmanPath, has, resolveModule, score } from './core/funcs'
-import Keymaps, { LocalMode, MapMode } from './core/keymaps'
+import Keymaps, { KeymapCallback, LocalMode, MapMode } from './core/keymaps'
 import * as ui from './core/ui'
 import Watchers from './core/watchers'
 import WorkspaceFolderController from './core/workspaceFolder'
@@ -255,7 +255,7 @@ export class Workspace {
   }
 
   public fixWin32unixFilepath(filepath: string): string {
-    return this.documentsManager.fixUnixPrefix(filepath, this.env.unixPrefix)
+    return this.documentsManager.fixUnixPrefix(filepath)
   }
 
   public checkPatterns(patterns: string[], folders?: WorkspaceFolder[]): Promise<boolean> {
@@ -311,12 +311,14 @@ export class Workspace {
   /**
    * Register autocmd on vim.
    */
-  public registerAutocmd(autocmd: Autocmd): Disposable {
+  public registerAutocmd(autocmd: Autocmd, disposables?: Disposable[]): Disposable {
     if (autocmd.request && autocmd.event !== 'BufWritePre') {
       let name = parseExtensionName(Error().stack)
       logger.warn(`Extension "${name}" registered synchronized autocmd "${autocmd.event}", which could be slow.`)
     }
-    return this.autocmds.registerAutocmd(autocmd)
+    let disposable = this.autocmds.registerAutocmd(autocmd)
+    if (disposables) disposables.push(disposable)
+    return disposable
   }
 
   /**
@@ -337,7 +339,7 @@ export class Workspace {
   /**
    * Check if selector match document.
    */
-  public match(selector: DocumentSelector, document: TextDocumentMatch): number {
+  public match(selector: DocumentSelector | DocumentFilter | string, document: TextDocumentMatch): number {
     return score(selector, document.uri, document.languageId)
   }
 
@@ -487,15 +489,15 @@ export class Workspace {
     return this.contentProvider.registerTextDocumentContentProvider(scheme, provider)
   }
 
-  public registerKeymap(modes: MapMode[], key: string, fn: Function, opts: Partial<KeymapOption> = {}): Disposable {
+  public registerKeymap(modes: MapMode[], key: string, fn: KeymapCallback, opts: Partial<KeymapOption> = {}): Disposable {
     return this.keymaps.registerKeymap(modes, key, fn, opts)
   }
 
-  public registerExprKeymap(mode: 'i' | 'n' | 'v' | 's' | 'x', key: string, fn: Function, buffer = false, cancel = true): Disposable {
+  public registerExprKeymap(mode: 'i' | 'n' | 'v' | 's' | 'x', key: string, fn: KeymapCallback, buffer = false, cancel = true): Disposable {
     return this.keymaps.registerExprKeymap(mode, key, fn, buffer, cancel)
   }
 
-  public registerLocalKeymap(bufnr: number, mode: LocalMode, key: string, fn: Function, notify = false): Disposable {
+  public registerLocalKeymap(bufnr: number, mode: LocalMode, key: string, fn: KeymapCallback, notify = false): Disposable {
     if (typeof arguments[0] === 'string') {
       bufnr = this.bufnr
       mode = arguments[0] as LocalMode